requirements.txt
aiogram>=3.0.0
python-dotenv>=0.19.0
aiosqlite>=0.17.0
pytz>=2021.3 